Move param handling helpers to separate file
